fix: login rerendering on login page
