+++ src/components/login/Login.tsx
@@ -14,14 +14,11 @@
  * limitations under the License.
  */
 
-import React, { useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { Switch, Redirect, Route, useLocation, useHistory } from 'react-router-dom'
 import {
     getCookie,
-    ServerErrors,
     Host,
-    showError,
-    CustomInput,
     URLS as CommonURL,
     ToastVariantType,
     ToastManager,
@@ -30,8 +27,6 @@ import {
     ButtonComponentType,
     ButtonVariantType,
     ButtonStyleType,
-    useUserEmail,
-    PasswordField,
     useSearchString,
     getComponentSpecificThemeClass,
     AppThemeType,
@@ -40,15 +35,13 @@ import {
     Icon,
 } from '@devtron-labs/devtron-fe-common-lib'
 import { importComponentFromFELibrary } from '@Components/common'
-import { ReactComponent as Help } from '@Icons/ic-help-outline.svg'
-import { REQUIRED_FIELD_MSG } from '@Config/constantMessaging'
-import { URLS, DOCUMENTATION, TOKEN_COOKIE_NAME } from '../../config'
-import { loginAsAdmin } from './login.service'
+import { URLS, TOKEN_COOKIE_NAME } from '../../config'
 import { dashboardAccessed } from '../../services/service'
 import './login.scss'
 import { getSSOConfigList } from '../../Pages/GlobalConfigurations/Authorization/SSOLoginServices/service'
 import { SSOConfigLoginList } from './login.types'
 import { SSOProvider } from './constants'
+import { LoginForm } from './LoginForm'
 
 const NetworkStatusInterface = !importComponentFromFELibrary('NetworkStatusInterface', null, 'function')
 
@@ -57,26 +50,10 @@ const getTermsAndConditions = importComponentFromFELibrary('getTermsAndCondition
 const Login = () => {
     const [continueUrl, setContinueUrl] = useState('')
     const [loginList, setLoginList] = useState<SSOConfigLoginList[]>([])
-    const [loading, setLoading] = useState(false)
-    const [form, setForm] = useState({
-        username: 'admin',
-        password: '',
-    })
-    const [errorMessage, setErrorMessage] = useState({
-        username: {
-            message: '',
-            isValid: true,
-        },
-        password: {
-            message: '',
-            isValid: true,
-        },
-    })
 
     const { searchParams } = useSearchString()
     const location = useLocation()
     const history = useHistory()
-    const { setEmail } = useUserEmail()
 
     const [initLoading, initResult] = useAsync(() => Promise.allSettled([getSSOConfigList(), dashboardAccessed()]), [])
 
@@ -134,58 +111,6 @@ const Login = () => {
         }
     }, [initLoading, initResult])
 
-    const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        e.persist()
-
-        const { name, value } = e.target
-        setErrorMessage((prevState) => ({
-            ...prevState,
-            [name]: {
-                message: value ? '' : REQUIRED_FIELD_MSG,
-                isValid: !!value,
-            },
-        }))
-
-        setForm({
-            ...form,
-            [name]: value,
-        })
-    }
-
-    const getDefaultRedirectionURL = () => {
-        const queryString = location.search.split('continue=')[1]
-        if (queryString) {
-            return queryString
-        }
-
-        if (!window._env_.HIDE_NETWORK_STATUS_INTERFACE && !!NetworkStatusInterface) {
-            return CommonURL.NETWORK_STATUS_INTERFACE
-        }
-
-        // NOTE: we don't have serverMode therefore defaulting to flag value
-        return window._env_.FEATURE_DEFAULT_LANDING_RB_ENABLE ? URLS.RESOURCE_BROWSER : URLS.APP
-    }
-
-    const onSubmitLogin = (e): void => {
-        e.preventDefault()
-        const data = form
-        setLoading(true)
-        loginAsAdmin(data)
-            .then((response) => {
-                if (response.result.token) {
-                    setLoading(false)
-                    const url = getDefaultRedirectionURL()
-                    setEmail(data.username)
-                    history.push(url)
-                    localStorage.setItem('isAdminLogin', 'true')
-                }
-            })
-            .catch((errors: ServerErrors) => {
-                showError(errors)
-                setLoading(false)
-            })
-    }
-
     const onClickSSO = () => {
         if (typeof Storage !== 'undefined') {
             localStorage.setItem('isSSOLogin', 'true')
@@ -228,76 +153,7 @@ const Login = () => {
         </div>
     )
 
-    const renderAdminLoginPage = () => {
-        const { search } = location
-
-        return (
-            <form className="flexbox-col dc__gap-32" autoComplete="on" onSubmit={onSubmitLogin} noValidate>
-                <div className="flexbox-col dc__gap-16">
-                    <CustomInput
-                        placeholder="Enter username"
-                        value={form.username}
-                        name="username"
-                        onChange={handleChange}
-                        label="User ID"
-                        required
-                        error={errorMessage.username.message}
-                    />
-                    <div className="flexbox-col dc__gap-4">
-                        <PasswordField
-                            placeholder="Enter password"
-                            value={form.password}
-                            name="password"
-                            onChange={handleChange}
-                            label="Password"
-                            required
-                            shouldShowDefaultPlaceholderOnBlur={false}
-                            autoFocus
-                            error={errorMessage.password.message}
-                        />
-
-                        <div className="flex left dc__gap-4">
-                            <Help className="fcb-5 icon-dim-16" />
-
-                            <a
-                                className="anchor fs-11 cb-5 lh-20"
-                                rel="noreferrer noopener"
-                                target="_blank"
-                                href={DOCUMENTATION.ADMIN_PASSWORD}
-                            >
-                                What is my admin password?
-                            </a>
-                        </div>
-                    </div>
-                </div>
-                <div className="flexbox-col dc__gap-12">
-                    <Button
-                        disabled={loading}
-                        isLoading={loading}
-                        dataTestId="login-button"
-                        text="Login"
-                        fullWidth
-                        size={ComponentSizeType.xl}
-                        buttonProps={{
-                            type: 'submit',
-                        }}
-                    />
-
-                    {loginList.length > 0 && (
-                        <Button
-                            dataTestId="sso-login"
-                            text="Login using SSO service"
-                            component={ButtonComponentType.link}
-                            linkProps={{
-                                to: `${URLS.LOGIN_SSO}${search}`,
-                            }}
-                            variant={ButtonVariantType.text}
-                        />
-                    )}
-                </div>
-            </form>
-        )
-    }
+    const renderAdminLoginPage = () => <LoginForm loginList={loginList} />
 
     const renderDevtronLogo = () => (
         <div className="flex column dc__gap-16 dc__text-center">
