+++ src/components/login/login.scss
@@ -68,51 +68,6 @@
     font-size: 13px;
 }
 
-.text-input {
-    border-radius: 4px;
-    border: solid 1px var(--N200);
-    padding: 12px 16px;
-    font-weight: 400;
-    font-size: 12px;
-
-    &:focus {
-        outline: none;
-        border: solid 1px var(--N700);
-    }
-
-    &::-webkit-input-placeholder {
-        font-size: 14px;
-        color: var(--N400);
-    }
-
-    &:-ms-input-placeholder {
-        font-size: 14px;
-        color: var(--N400);
-    }
-
-    &::placeholder {
-        font-size: 14px;
-        color: var(--N400);
-    }
-}
-
-.text-input--username {
-    width: 100%;
-    height: 44px;
-    margin-bottom: 24px;
-}
-
-.text-input--pwd {
-    width: 100%;
-    height: 44px;
-}
-
-@media (min-width: 1440px) {
-    .text-input {
-        margin-bottom: 32px;
-    }
-}
-
 .sso-icons {
     position: relative;
     border-radius: 4px;
@@ -139,18 +94,4 @@
         border-color: 1px solid var(--B500);
         background-color: var(--B500);
     }
-}
-
-.login__text {
-    width: 300px;
-    font-size: 16px;
-    font-weight: normal;
-    font-style: normal;
-    font-stretch: normal;
-    line-height: 1.5;
-    letter-spacing: normal;
-    text-align: center;
-    color: var(--black);
-    margin: auto;
-    margin-bottom: 48px;
 }
